refactor: extract sidebar rendering into ui/sidebar.rs (#357)
First slice of the ui.rs decomposition: move ui.rs to ui/mod.rs and
peel draw_sidebar (~135 lines) into a dedicated submodule. Sidebar
rendering had no cross-cutting state worth keeping inline -- it reads
conversation order, mute/blocked decorations, and the sidebar filter
overlay, then writes the inner Rect back to app.mouse.sidebar_inner
for click-to-focus.

Snapshot files moved from src/snapshots/ to src/ui/snapshots/ to
follow insta's source-relative path resolution after the directory
rename.

ui/mod.rs: 5272 -> 5140 lines.
